% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Mmd.R
\name{compute_Mmd}
\alias{compute_Mmd}
\title{Computes the Mmd matrix using Spatial Deformation MCMC output}
\usage{
compute_Mmd(MCMC_output, response, sites, Map, dd)
}
\arguments{
\item{MCMC_output}{A list containing the output of \code{SpatialDeformationMCMC}, which includes the matrix \code{MDef}.}

\item{response}{A matrix of observed values for the response variable.}

\item{sites}{A matrix containing the geographic coordinates of the locations.}

\item{Map}{A matrix representing the geographic coordinates of the map boundaries.}

\item{dd}{The number of divisions for the grid.}
}
\value{
Mmd A matrix with computed deformations based on the MCMC output.
}
\description{
Computes the Mmd matrix using Spatial Deformation MCMC output
}
