% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GT.R
\docType{data}
\name{GT}
\alias{GT}
\title{Transition Matrix for State-Space Model at Time t}
\format{
A matrix with \code{p} rows and \code{p} columns:
\describe{
\item{p}{The number of state variables in the model.}
}
}
\usage{
GT
}
\description{
The \code{GT} matrix is the transition matrix used in the state-space model at a given time \code{t}. This matrix
describes the evolution of the state variables over time, allowing the model to capture the temporal
dependencies and transitions between states in the spatiotemporal process.
}
\details{
The \code{GT} matrix serves as the transition matrix in the state-space model at time \code{t}. It defines how the
state variables at time \code{t - 1} influence the state variables at time \code{t}, playing a crucial role in
modeling the temporal dynamics of the system. This matrix is essential for capturing the evolution and
propagation of states across time steps in the model.
}
\examples{
# Load the GT matrix
data(GT)

# Display the structure of GT
str(GT)

}
\keyword{datasets}
