% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map.R
\docType{data}
\name{Map}
\alias{Map}
\title{Border Map of Southern Brazil}
\format{
A data frame with \code{m} rows and 2 variables:
\describe{
\item{longitude}{A numeric vector representing the longitude of each point on the border outline, in decimal degrees.}
\item{latitude}{A numeric vector representing the latitude of each point on the border outline, in decimal degrees.}
}
}
\source{
\url{https://portal.inmet.gov.br/} (National Institute of Meteorology, Brazil)
}
\usage{
Map
}
\description{
The \code{Map} dataset provides the geographic coordinates outlining the border of the Southern region of Brazil.
This dataset is useful for spatial visualizations and mapping purposes, especially in analyses focusing on
this specific geographic area.
}
\details{
The \code{Map} dataset captures the geographic boundary of the Southern region of Brazil, defined by a series
of latitude and longitude coordinates that trace the contour of the region. This data can be used as a
background layer for mapping applications or as a reference for spatial analyses within this region.
}
\examples{
# Load the Map dataset
data(Map)

# Plot the map outline, assuming Map is a matrix
plot(Map[,1], Map[,2], type = "l",
     main = "Border Map of Southern Brazil",
     xlab = "Longitude", ylab = "Latitude")

}
\keyword{datasets}
