% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_interpolation.R
\name{spatial_interpolation}
\alias{spatial_interpolation}
\title{Spatial Interpolation based on MCMC Spatiotemporal Model Output}
\usage{
spatial_interpolation(
  MCMC_output,
  response,
  FT,
  MatFFT,
  GT,
  sites,
  sitesNO,
  indi,
  FTNO,
  MAP,
  bd,
  dd = 11
)
}
\arguments{
\item{MCMC_output}{A list containing the output of \code{SpatialDeformationMCMC}, including \code{MDef}, \code{MTheta}, \code{Msigmak}, \code{Mkappa}, and \code{MPhi}.}

\item{response}{A matrix of observed values for the response variable.}

\item{FT}{A matrix of covariates for the observed locations.}

\item{MatFFT}{A covariate matrix across time and space.}

\item{GT}{An evolution matrix.}

\item{sites}{A matrix containing the coordinates of observed locations.}

\item{sitesNO}{A matrix containing the coordinates of non-monitored locations to interpolate.}

\item{indi}{A vector of indices representing subsets of sites.}

\item{FTNO}{A matrix of covariates for the non-monitored locations.}

\item{MAP}{A matrix representing the map area.}

\item{bd}{A scalar parameter for the spatial correlation function.}

\item{dd}{Integer indicating the grid resolution (default: 11).}
}
\value{
A matrix, \code{MatvecNO}, containing the interpolated values at the non-monitored locations for each time step.
}
\description{
This function performs spatial interpolation for non-monitored locations based on the output of the \code{SpatialDeformationMCMC} model.
}
