% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stdyndef-package.R
\docType{package}
\name{Stdyndef-package}
\alias{Stdyndef}
\alias{Stdyndef-package}
\title{Stdyndef: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Fidel Morales \email{fidel.ec@ufrn.br} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
