% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stdyndef.R
\name{summary_stdyndef}
\alias{summary_stdyndef}
\title{Summary Function for SpatialDeformationMCMC Output}
\usage{
summary_stdyndef(MCMC_output)
}
\arguments{
\item{MCMC_output}{A list containing the matrices Msigmak, MPhi, and Mkappa, which are
the outputs from the SpatialDeformationMCMC function.}
}
\value{
A data frame with summary statistics (mean, 2.5\%, and 97.5\%) for each parameter.
}
\description{
This function takes the output of the SpatialDeformationMCMC function and returns
a summary table with mean, 2.5\% quantile, and 97.5\% quantile for sigmak, phik, and pik.
}
