% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIC.R
\name{DIC}
\alias{DIC}
\title{Title}
\usage{
DIC(output, response, FT, MatFFT, GT, sites, GAMA)
}
\arguments{
\item{output}{a value}

\item{response}{a value}

\item{FT}{a value}

\item{MatFFT}{a value}

\item{GT}{a value}

\item{sites}{a value}

\item{GAMA}{a value}
}
\description{
Title
}
