% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialDeformationMCMC.R
\name{SpatialDeformationMCMC}
\alias{SpatialDeformationMCMC}
\title{Title}
\usage{
SpatialDeformationMCMC(
  response,
  FT,
  MatFFT,
  GT,
  sites,
  GAMA,
  prior = list(m0 = as.matrix(rep(0, nrow(FT))), mvarphi = 1/18, C0 = diag(1000,
    nrow(FT)), S0 = diag(0.01, nrow(response)), n0 = 1, n0est = n0 + nrow(response),
    asigmad = 10002, bsigmad = 10001, asigmala = 0.01, bsigmala = 0.01, aepsilon =
    rep(0.01, nrow(GAMA)), axi = rep(0.01, nrow(GAMA)), bepsilon = rep(0.01, nrow(GAMA)),
    bxi = rep(0.01, nrow(GAMA)), aaphi = rep(0.01, nrow(GAMA)), bbphi = rep(0.01,
    nrow(GAMA)), u1phi = rep(500, nrow(GAMA))),
  iteration,
  burnin,
  jump
)
}
\arguments{
\item{response}{a value}

\item{FT}{a value}

\item{MatFFT}{a value}

\item{GT}{a value}

\item{sites}{a value}

\item{GAMA}{a value}

\item{prior}{a value}

\item{iteration}{a value}

\item{burnin}{a value}

\item{jump}{a value}
}
\value{
a matrix
}
\description{
Title
}
