% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatio_temporal_interpolation.R
\name{spatio_temporal_interpolation}
\alias{spatio_temporal_interpolation}
\title{Spatio-Temporal Interpolation for Response Variable}
\usage{
spatio_temporal_interpolation(
  MCMC_output,
  response,
  FT,
  MatFFT,
  GT,
  sites,
  INDI,
  GAMA,
  FTNO,
  RS
)
}
\arguments{
\item{MCMC_output}{A list containing the output matrices Msigmak, MPhi, Mkappa, MTheta, MDef, and the parameter bd from SpatialDeformationMCMC.}

\item{response}{Matrix of observed values for the response variable.}

\item{FT}{Covariate matrix for observed locations.}

\item{MatFFT}{Covariate matrix for all time-space points.}

\item{GT}{Evolution matrix.}

\item{sites}{Matrix of observed locations coordinates.}

\item{INDI}{Indices of the grid points for interpolation.}

\item{GAMA}{Matrix defining subsets of locations for interpolation.}

\item{FTNO}{Covariate matrix for non-observed locations.}

\item{RS}{Matrix representing the map area for spatial interpolation.}
}
\value{
A matrix \code{MatvecNO} containing interpolated values for non-observed locations for each time step.
}
\description{
This function performs spatio-temporal interpolation for a response variable based on the output
of the SpatialDeformationMCMC model.
}
